+++ api/controllers/login.js
@@ -3,6 +3,7 @@
 const https = require('https')
 const jwt = require('jsonwebtoken')
 const logger = require('../../config/logger')
+const util = require('../util/util')
 
 const get = (request, response) => {
   request.body = {
@@ -16,20 +17,20 @@ const post = (request, response) => {
   var serviceTicket = request.body['ticket']
 
   if (!serviceTicket) {
-    buildUnauthorized(response)
+    util.buildUnauthorizedResponse(response)
     return
   }
 
   validateTicket(serviceTicket, (data) => {
     if (!data) {
-      buildInternalErrorResponse(response)
+      util.buildInternalErrorResponse(response)
       return
     }
 
     var json = JSON.parse(data)
     if (json.serviceResponse.authenticationFailure) {
       logger.debug('Login failure: ' + json.serviceResponse.authenticationFailure.description)
-      buildUnauthorized(response)
+      util.buildUnauthorizedResponse(response)
       return
     }
     var guid = json.serviceResponse.authenticationSuccess.attributes.theKeyGuid[0].toLowerCase()
@@ -68,24 +69,10 @@ const buildJwt = (guid, response) => {
   jwt.sign({ guid: guid }, process.env.JWT_SECRET, { expiresIn: '1h' }, (error, token) => {
     if (error) {
       logger.error(error)
-      buildInternalErrorResponse(response)
+      util.buildInternalErrorResponse(response)
     } else {
       response.json(token)
     }
-  })
-}
-
-const buildUnauthorized = (response) => {
-  response.status(401)
-  response.json({
-    message: 'Unauthorized'
-  })
-}
-
-const buildInternalErrorResponse = (response) => {
-  response.status(500)
-  response.json({
-    message: 'Internal Server Error'
   })
 }
